Remove XSDK_USE_COMPILER_ENV_VARS and make always true
During the debate about what the XSDK default behavior should be, we were
discussing how to implement the desired behavior where compilers and flags
would *not* be found in the env by default.  I had experimented with this in
the XSDKDefaults.cmake module and came to the conclusion that if you wanted to
disable CMake behavior where you disabled finding compilers in the env then
you had to do the default compiler search yourself.  This would have made the
XSDK default compiler search different than the built-in CMake compiler
search.  That would have lead to a lot of confusion by users.

The new refactored code and tests eliminate XSDK_USE_COMPILER_ENV_VARS and
instead replace it with TRUE.  What that does is to cause the code in
XSDKDefaults.cmake to find the compilers and flag vars FC CC, CXX, FFLAGS,
FCFLAGS, CFLAGS, CXXFLAGS in the env, and then just print them out and set the
CMake vars.  This is much better than what raw CMake does which is to be
silent.